CRUD functions are now protected by admin except READ

+++ application/controllers/Blog.php
@@ -9,6 +9,10 @@ class Blog extends CI_Controller {
   }
 
   public function index() {
+    $data['is_admin'] = FALSE;
+    if ($this->session->usertype === 'admin'){
+    $data['is_admin'] = TRUE;
+    }
     $this->load->model('user_model');
     $data['blog'] = $this->blog_model->get_article();
     $data['title'] = 'Blog index';
@@ -29,6 +33,9 @@ class Blog extends CI_Controller {
   }
 
   public function create() {
+    if(!$this->check_permissions('admin')) {
+      redirect(base_url().'blog/');
+    }
     $data['title'] = 'Create a blog item';
     $slug = url_title($this->input->post('title'), 'dash', TRUE);
     $this->form_validation->set_rules('title', 'Title', 'required');
@@ -45,11 +52,17 @@ class Blog extends CI_Controller {
   }
 
   public function delete($id) {
+    if(!$this->check_permissions('admin')) {
+      redirect(base_url().'blog/');
+    }
     $this->blog_model->delete_article($id);
     $this->load->view('blog/delete_success');
   }
 
   public function update($id) {
+    if(!$this->check_permissions('admin')) {
+      redirect(base_url().'blog/');
+    }
     $data['blog_item'] = $this->blog_model->get_article_by_id($id);
     $data['title'] = 'Edit a blog item';
     $this->form_validation->set_rules('title', 'Title', 'required');
@@ -62,6 +75,15 @@ class Blog extends CI_Controller {
     else {
       $this->blog_model->set_article();
       $this->load->view('blog/update_success');
+    }
+  }
+
+  function check_permissions($required) {
+    $usertype = $this->session->userdata('usertype');
+    if($required == 'admin') {
+      if($usertype == 'admin') {
+        return TRUE;
+      }
     }
   }
 }
